Simplify energy offset and fix Cantera transport
Removed the attempt at directly modifying the thermodynamic polynomials
as it turned out to be internally inconsistent, messing up equilibrium
calculations and constants. Replaced it with a per-species energy offset
which is removed before passing energy to the Cantera routines.

Added THERMO_TYPE parameter to thermodynamic classes, which was needed
by the Cantera transport. Also added some missing broadcasting of array
shapes to ensure consistency.