Generalize explicit Runge-Kutta time steppers
A new base class for Runge-Kutta time integrators was added,
generalizing the standard explicit format. The 3- and 4-stage SSPRK3
algorithms were added using this format. It should be straightforward
to add the 4-stage RK4 version to this as well.

Low-storage Runge-Kutta time integrators could also be generalized in a
similar fashion.